# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Network School OAuth Provider",
    "description": "OAuth 2.0 / OpenID Connect identity provider for Network School.\n\nLets third-party apps request NS user data via standard OAuth scopes \u2014 the same pattern as \"Sign in with Google.\"\n\n## Flows\n\n- **Authorization Code + PKCE** \u2014 user-facing \"Sign in with Network School\" flow\n- **Client Credentials** \u2014 machine-to-machine, no user context\n\n## Scopes\n\n| Scope | Claims |\n|-------|--------|\n| `openid` | `sub` |\n| `email` | `email`, `email_verified` |\n| `profile` | `name`, `picture` |\n| `roles` | `roles` (Discord roles, live) |\n| `date_joined` | `date_joined` |\n| `offline_access` | refresh tokens |\n\n## Authentication\n\n- **User auth**: Discord OAuth2 (NS guild membership required)\n- **Sessions**: HS256 JWT in httponly cookie (`ns_session`)\n- **Access tokens**: RS256 JWT (auto-generated RSA keys)\n- **Client auth**: `client_secret` verified via bcrypt\n",
    "license": {
      "name": "Private",
      "url": "https://ns.com/"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/discord": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Redirect to Discord OAuth2 login",
        "description": "Generates a CSRF state token, stores the `next` URL, and redirects the user to Discord's OAuth2 authorization page. After Discord login, the user is sent to `/auth/discord/callback`.",
        "operationId": "discord_login_auth_discord_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "302": {
            "description": "Redirect to Discord OAuth2 authorization page."
          }
        }
      }
    },
    "/auth/discord/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Discord OAuth2 callback",
        "description": "Handles the Discord OAuth2 callback. Validates state, exchanges the authorization code for tokens, fetches the Discord user profile, verifies NS guild membership (rejects non-members), fetches guild roles, creates/updates the user, sets the `ns_session` cookie, and redirects to the stored `next` URL.",
        "operationId": "discord_callback_auth_discord_callback_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "302": {
            "description": "Redirect to `next` URL with session cookie set."
          }
        }
      }
    },
    "/auth/session/exchange": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange a session relay code for a session token",
        "description": "Accepts a single-use code (from the Discord login redirect) and returns a session JWT. The code expires after 60 seconds and can only be used once.",
        "operationId": "exchange_session_code_auth_session_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCodeExchange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session token returned.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "`invalid_code` \u2014 code missing, expired, or already used."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get current user",
        "description": "Returns the authenticated user's profile from the `ns_session` cookie. Used by the consent page to display who is granting access.",
        "operationId": "get_current_user_auth_me_get",
        "responses": {
          "200": {
            "description": "User profile.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "`not_authenticated` (no/invalid session) or `user_not_found`."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionCodeExchange": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "SessionCodeExchange"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}